[REFACTOR] 디렉토리 구조를 컨벤션에 맞게 수정
service(dir) -> application,
repository를 domain 내로 이동
